Match ProjectileUpdate
Not sure how else to get in touch, so feel free to close this, and fix it properly. But I found a match for this function while just browsing the code. (Really helpful including the decomp.me link in here by the way). My matching decomp.me is here: https://decomp.me/scratch/woOst